+++ NEWS.md
@@ -4,6 +4,8 @@
 
 ## Fixes since v1.3.11
 - filename of the annotBEBover0.9 fasta file now contains parameters used (as different parameters will give different values)
+- added addGeneName option to `pw_combineParsedOutfilesWide.pl` and `pw_wideFormatCombineUnmaskedAndMaskedResults.pl` but default behavior is unchanged. If you supply the `--genename=1` option, those scripts will try to add gene name as the first column (extracts it from filename, first field before `_`).
+
 
 ## paml_wrapper_v1.3.11
 - docker container's R installation failed to install the ape package, so I got an error when trying to plot the trees. I think the newer ape package version available at CRAN probably fails to install on the very old version of R I have in the container. I fixed it by specifying an older version of ape, and by manually installing the Rcpp package rather than doing it as a dependency.